Move utxo types into 'types' module
